hotfix for pdfpage gather layouts

+++ logger.py
@@ -225,7 +225,7 @@ class Log(logging.Logger):
         if GlobalLogContext.context_current.heading is not False:
             context_msg_level = GlobalLogContext.context_current.heading_level if type(GlobalLogContext.context_current.heading_level) is int else logging.INFO
             if DEBUG_FLAG.LOGGER_CONTEXT_OPENING is True: print(f"{'@'*20} logger.logContextStatusOpening @ context msg level={context_msg_level} {'@'*20}")
-            logging.Logger.logf(self, int(context_msg_level), None, {'context': LogContextStatus.OPENING, 'heading': f"{GlobalLogContext.context_current.getHeading()}" } )
+            logging.Logger.log(self, int(context_msg_level), None, {'context': LogContextStatus.OPENING, 'heading': f"{GlobalLogContext.context_current.getHeading()}" } )
         GlobalLogContext.status = LogContextStatus.CURRENT
         GlobalLogContext.context_current.context_count = GlobalLogContext.context_current.context_count + 1
         # print("end of logContextStatusOpening()")
@@ -241,7 +241,7 @@ class Log(logging.Logger):
             if DEBUG_FLAG.LOGGER_CONTEXT_CLOSING is True: print(f"{'@'*20} logger.logContextStatusClosing({GlobalLogContext.context_current.getHeading()}) @ level={context_exit_level}(prior level), new level={self.level} {'@'*20}")
             save_level = self.level
             self.level = context_exit_level
-            logging.Logger.logf(self, int(GlobalLogContext.context_current.heading_level), None, {'context': LogContextStatus.CLOSING, 'heading': GlobalLogContext.context_current.getHeading()} )
+            logging.Logger.log(self, int(GlobalLogContext.context_current.heading_level), None, {'context': LogContextStatus.CLOSING, 'heading': GlobalLogContext.context_current.getHeading()} )
             self.level = save_level
         if DEBUG_FLAG.LOGGER_CONTEXT_CLOSING is True: print(f"{'@'*20} logger.logContextStatusClosing - context set to NOCONTEXT")
         GlobalLogContext.status = LogContextStatus.NOCONTEXT
